// src/generated-api/src/models/DomainCreateInvitationRequest.ts
export interface DomainCreateInvitationRequest {
    /**
     * 
     * @type {number}
     * @memberof DomainCreateInvitationRequest
     */
    advertiserOrgId: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof DomainCreateInvitationRequest
     */
    allowedAffiliateOrgIds?: Array<number>;
    /**
     * 
     * @type {boolean}
     * @memberof DomainCreateInvitationRequest
     */
    defaultAllAffiliatesVisible?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof DomainCreateInvitationRequest
     */
    defaultAllCampaignsVisible?: boolean;
    /**
     * 
     * @type {Array<number>}
     * @memberof DomainCreateInvitationRequest
     */
    defaultVisibleAffiliateIds?: Array<number>;
    /**
     * 
     * @type {Array<number>}
     * @memberof DomainCreateInvitationRequest
     */
    defaultVisibleCampaignIds?: Array<number>;
    /**
     * 
     * @type {string}
     * @memberof DomainCreateInvitationRequest
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof DomainCreateInvitationRequest
     */
    expiresAt?: string;
    /**
     * 
     * @type {number}
     * @memberof DomainCreateInvitationRequest
     */
    maxUses?: number;
    /**
     * 
     * @type {string}
     * @memberof DomainCreateInvitationRequest
     */
    message?: string;
    /**
     * 
     * @type {string}
     * @memberof DomainCreateInvitationRequest
     */
    name: string;
}