// src/generated-api/src/models/DomainUpdateInvitationRequest.ts
export interface DomainUpdateInvitationRequest {
    /**
     * 
     * @type {Array<number>}
     * @memberof DomainUpdateInvitationRequest
     */
    allowedAffiliateOrgIds?: Array<number>;
    /**
     * 
     * @type {boolean}
     * @memberof DomainUpdateInvitationRequest
     */
    defaultAllAffiliatesVisible?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof DomainUpdateInvitationRequest
     */
    defaultAllCampaignsVisible?: boolean;
    /**
     * 
     * @type {Array<number>}
     * @memberof DomainUpdateInvitationRequest
     */
    defaultVisibleAffiliateIds?: Array<number>;
    /**
     * 
     * @type {Array<number>}
     * @memberof DomainUpdateInvitationRequest
     */
    defaultVisibleCampaignIds?: Array<number>;
    /**
     * 
     * @type {string}
     * @memberof DomainUpdateInvitationRequest
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof DomainUpdateInvitationRequest
     */
    expiresAt?: string;
    /**
     * 
     * @type {number}
     * @memberof DomainUpdateInvitationRequest
     */
    maxUses?: number;
    /**
     * 
     * @type {string}
     * @memberof DomainUpdateInvitationRequest
     */
    message?: string;
    /**
     * 
     * @type {string}
     * @memberof DomainUpdateInvitationRequest
     */
    name?: string;
    /**
     * 
     * @type {DomainInvitationStatus}
     * @memberof DomainUpdateInvitationRequest
     */
    status?: DomainInvitationStatus;
}